ml/requirements.txt
azureml-sdk==1.15.0
azureml-mlflow==1.15.0
azureml-defaults==1.15.0
scikit-learn==0.23.2
pandas==1.1.3
joblib==0.16.0
python-dotenv==0.14.0
imblearn==0.0
xgboost==1.3.3
matplotlib==3.2.1
seaborn==0.11.0